_sdist: split SdistBuilder into its own module
